+++ pages/_app.js
@@ -5,7 +5,7 @@ import { AppProvider } from "@context/AppContext";
 export default function App({ Component, pageProps }) {
   return (
     <AppProvider>
-      <Component {...pageProps} />;
+      <Component {...pageProps} />
     </AppProvider>
   );
 }