+++ CHANGELOG.md
@@ -4,6 +4,8 @@ NextJS Starter is a boilerplate project that serves as a starting point for all 
 
 ## Upcoming Release: v1.0.0
 
+- [#6](https://github.com/jtcarlos/nextjs-starter/issues/6) - app(feat): add page layout component
+- [#5](https://github.com/jtcarlos/nextjs-starter/issues/5) - app(feat): add helmet component
 - [#3](https://github.com/jtcarlos/nextjs-starter/issues/3) - app(feat): update folder structure
 - [#2](https://github.com/jtcarlos/nextjs-starter/issues/2) - app(feat): add tailwind css
 - [#1](https://github.com/jtcarlos/nextjs-starter/issues/1) - app(feat): initialize nextjs application